fix more signedness warnings

+++ Community/General/Projects/PDF3DFigurePage/Sources/PDF3DFigurePage_Utils/PDF3DFigurePage_Utils_Lists.cpp
@@ -26,7 +26,7 @@ void PDF3DFigurePage_Utils::_getMaxTypeIDFromList(ml::XMarkerList& list, int& ma
 {
   size_t listLength = list.size();
 
-  for (int i = 0; i < listLength; i++)
+  for (size_t i = 0; i < listLength; i++)
   {
     ml::XMarker thisListEntry = list[i];
 
@@ -43,7 +43,7 @@ void PDF3DFigurePage_Utils::_getMaxTypeIDFromList(ml::IndexPairList& list, int& 
 {
   size_t listLength = list.size();
 
-  for (int i = 0; i < listLength; i++)
+  for (size_t i = 0; i < listLength; i++)
   {
     ml::IndexPair thisListEntry = list[i];
